Use the public yjs server
