+++ playground/package.json
@@ -48,6 +48,7 @@
     "vue": "^3.4.35",
     "vue-tsc": "^2.0.29",
     "y-protocols": "^1.0.6",
+    "y-websocket": "^2.0.4",
     "yjs": "^13.6.18"
   }
 }
